# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
    "/products" : {
      "get" : {
        "summary" : "상품 목록을 조회합니다.",
        "description" : "상품 목록을 조회합니다.",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "query",
          "description" : "카테고리별 조회를 위한 카테고리 ID",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "page" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "size" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "sort" : {
                    "type" : "array",
                    "nullable" : false,
                    "items" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "required" : false
        },
        "responses" : {
          "200" : {
            "description" : "상품 요약 정보 목록",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "content" : {
                      "type" : "array",
                      "description" : "현재 페이지의 데이터 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "상품 ID",
                            "format" : "int64"
                          },
                          "name" : {
                            "type" : "string",
                            "description" : "상품명"
                          },
                          "startingPrice" : {
                            "type" : "number",
                            "description" : "상품 가격"
                          },
                          "productStatus" : {
                            "type" : "string",
                            "description" : "{ACTIVE=ACTIVE, INACTIVE=INACTIVE, SOLD_OUT=SOLD_OUT, AUCTION_COMPLETED=AUCTION_COMPLETED, DELETED=DELETED}",
                            "enum" : [ "ACTIVE", "INACTIVE", "SOLD_OUT", "AUCTION_COMPLETED", "DELETED" ]
                          },
                          "productCondition" : {
                            "type" : "string",
                            "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                            "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                          },
                          "listingType" : {
                            "type" : "string",
                            "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                            "enum" : [ "NORMAL", "AUCTION" ]
                          },
                          "sellerNickname" : {
                            "type" : "string",
                            "description" : "판매자 닉네임"
                          },
                          "categoryName" : {
                            "type" : "string",
                            "description" : "카테고리명"
                          },
                          "thumbnailUrl" : {
                            "type" : "string",
                            "description" : "대표 썸네일 이미지 URL"
                          },
                          "createdAt" : {
                            "type" : "string",
                            "description" : "상품 등록 일시"
                          }
                        },
                        "description" : "상품 목록 조회 응답 DTO.\n\n<p>상품 목록 화면에 필요한 요약 정보를 클라이언트에 반환합니다."
                      }
                    },
                    "page" : {
                      "type" : "integer",
                      "description" : "현재 페이지 번호 (0-based)"
                    },
                    "size" : {
                      "type" : "integer",
                      "description" : "페이지 크기"
                    },
                    "totalElements" : {
                      "type" : "integer",
                      "description" : "전체 데이터 건수",
                      "format" : "int64"
                    },
                    "totalPages" : {
                      "type" : "integer",
                      "description" : "전체 페이지 수"
                    },
                    "first" : {
                      "type" : "boolean",
                      "description" : "첫 번째 페이지 여부"
                    },
                    "last" : {
                      "type" : "boolean",
                      "description" : "마지막 페이지 여부"
                    }
                  },
                  "description" : "상품 요약 정보 목록"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "신규 상품을 등록합니다.",
        "description" : "신규 상품을 등록합니다.",
        "parameters" : [ {
          "name" : "nickname",
          "in" : "header",
          "description" : "요청한 판매자 닉네임",
          "required" : true,
          "schema" : { }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "description", "images", "listingType", "name", "productCondition", "startingPrice" ],
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "description" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "startingPrice" : {
                    "minimum" : 0.01,
                    "type" : "number",
                    "description" : ""
                  },
                  "productCondition" : {
                    "type" : "string",
                    "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                    "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                  },
                  "listingType" : {
                    "type" : "string",
                    "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                    "enum" : [ "NORMAL", "AUCTION" ]
                  },
                  "categoryId" : {
                    "type" : "integer",
                    "description" : "",
                    "format" : "int64"
                  },
                  "images" : {
                    "type" : "array",
                    "description" : "",
                    "items" : {
                      "required" : [ "imageType", "imageUrl", "sortOrder" ],
                      "type" : "object",
                      "properties" : {
                        "imageUrl" : {
                          "type" : "string",
                          "description" : "상품 이미지 URL"
                        },
                        "imageType" : {
                          "type" : "string",
                          "description" : "{THUMBNAIL=THUMBNAIL, DETAIL=DETAIL}",
                          "enum" : [ "THUMBNAIL", "DETAIL" ]
                        },
                        "sortOrder" : {
                          "minimum" : 0,
                          "type" : "integer",
                          "description" : "이미지 노출 순서"
                        }
                      },
                      "description" : "상품 이미지 요청 DTO.\n\n<p>상품 등록 및 수정 요청에서 이미지 URL과 정렬 정보를 함께 전달할 때 사용합니다."
                    }
                  }
                },
                "description" : "등록할 상품 정보 (images 포함)"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "등록된 상품 상세 정보 (HTTP 201)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "상품 ID",
                      "format" : "int64"
                    },
                    "name" : {
                      "type" : "string",
                      "description" : "상품명"
                    },
                    "description" : {
                      "type" : "string",
                      "description" : "상품 설명"
                    },
                    "price" : {
                      "type" : "number",
                      "description" : "상품 가격"
                    },
                    "productStatus" : {
                      "type" : "string",
                      "description" : "{ACTIVE=ACTIVE, INACTIVE=INACTIVE, SOLD_OUT=SOLD_OUT, AUCTION_COMPLETED=AUCTION_COMPLETED, DELETED=DELETED}",
                      "enum" : [ "ACTIVE", "INACTIVE", "SOLD_OUT", "AUCTION_COMPLETED", "DELETED" ]
                    },
                    "productCondition" : {
                      "type" : "string",
                      "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                      "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                    },
                    "listingType" : {
                      "type" : "string",
                      "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                      "enum" : [ "NORMAL", "AUCTION" ]
                    },
                    "sellerNickname" : {
                      "type" : "string",
                      "description" : "판매자 닉네임"
                    },
                    "categoryId" : {
                      "type" : "integer",
                      "description" : "카테고리 ID",
                      "format" : "int64"
                    },
                    "categoryName" : {
                      "type" : "string",
                      "description" : "카테고리명"
                    },
                    "images" : {
                      "type" : "array",
                      "description" : "상품 이미지 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "이미지 ID",
                            "format" : "int64"
                          },
                          "imageUrl" : {
                            "type" : "string",
                            "description" : "상품 이미지 URL"
                          },
                          "imageType" : {
                            "type" : "string",
                            "description" : "{THUMBNAIL=THUMBNAIL, DETAIL=DETAIL}",
                            "enum" : [ "THUMBNAIL", "DETAIL" ]
                          },
                          "sortOrder" : {
                            "type" : "integer",
                            "description" : "이미지 노출 순서"
                          }
                        },
                        "description" : "상품 이미지 응답 DTO.\n\n<p>상품 상세 정보에 포함되는 개별 이미지 메타데이터를 나타냅니다."
                      }
                    },
                    "createdAt" : {
                      "type" : "string",
                      "description" : "상품 생성 일시"
                    },
                    "updatedAt" : {
                      "type" : "string",
                      "description" : "상품 수정 일시"
                    }
                  },
                  "description" : "등록된 상품 상세 정보 (HTTP 201)"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}" : {
      "get" : {
        "summary" : "상품 상세 정보를 조회합니다.",
        "description" : "상품 상세 정보를 조회합니다.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "조회할 상품 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "상품 상세 정보",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "상품 ID",
                      "format" : "int64"
                    },
                    "name" : {
                      "type" : "string",
                      "description" : "상품명"
                    },
                    "description" : {
                      "type" : "string",
                      "description" : "상품 설명"
                    },
                    "price" : {
                      "type" : "number",
                      "description" : "상품 가격"
                    },
                    "productStatus" : {
                      "type" : "string",
                      "description" : "{ACTIVE=ACTIVE, INACTIVE=INACTIVE, SOLD_OUT=SOLD_OUT, AUCTION_COMPLETED=AUCTION_COMPLETED, DELETED=DELETED}",
                      "enum" : [ "ACTIVE", "INACTIVE", "SOLD_OUT", "AUCTION_COMPLETED", "DELETED" ]
                    },
                    "productCondition" : {
                      "type" : "string",
                      "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                      "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                    },
                    "listingType" : {
                      "type" : "string",
                      "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                      "enum" : [ "NORMAL", "AUCTION" ]
                    },
                    "sellerNickname" : {
                      "type" : "string",
                      "description" : "판매자 닉네임"
                    },
                    "categoryId" : {
                      "type" : "integer",
                      "description" : "카테고리 ID",
                      "format" : "int64"
                    },
                    "categoryName" : {
                      "type" : "string",
                      "description" : "카테고리명"
                    },
                    "images" : {
                      "type" : "array",
                      "description" : "상품 이미지 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "이미지 ID",
                            "format" : "int64"
                          },
                          "imageUrl" : {
                            "type" : "string",
                            "description" : "상품 이미지 URL"
                          },
                          "imageType" : {
                            "type" : "string",
                            "description" : "{THUMBNAIL=THUMBNAIL, DETAIL=DETAIL}",
                            "enum" : [ "THUMBNAIL", "DETAIL" ]
                          },
                          "sortOrder" : {
                            "type" : "integer",
                            "description" : "이미지 노출 순서"
                          }
                        },
                        "description" : "상품 이미지 응답 DTO.\n\n<p>상품 상세 정보에 포함되는 개별 이미지 메타데이터를 나타냅니다."
                      }
                    },
                    "createdAt" : {
                      "type" : "string",
                      "description" : "상품 생성 일시"
                    },
                    "updatedAt" : {
                      "type" : "string",
                      "description" : "상품 수정 일시"
                    }
                  },
                  "description" : "상품 상세 정보"
                }
              }
            }
          }
        }
      },
      "put" : {
        "summary" : "기존 상품 정보를 수정합니다.",
        "description" : "기존 상품 정보를 수정합니다.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "수정할 상품 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "nickname",
          "in" : "header",
          "description" : "요청한 판매자 닉네임",
          "required" : true,
          "schema" : { }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "categoryId", "description", "images", "listingType", "name", "productCondition", "productStatus", "startingPrice" ],
                "type" : "object",
                "properties" : {
                  "name" : {
                    "type" : "string",
                    "description" : "상품명"
                  },
                  "description" : {
                    "type" : "string",
                    "description" : "상품 설명"
                  },
                  "startingPrice" : {
                    "minimum" : 0.01,
                    "type" : "number",
                    "description" : "상품 가격"
                  },
                  "productStatus" : {
                    "type" : "string",
                    "description" : "{ACTIVE=ACTIVE, INACTIVE=INACTIVE, SOLD_OUT=SOLD_OUT, AUCTION_COMPLETED=AUCTION_COMPLETED, DELETED=DELETED}",
                    "enum" : [ "ACTIVE", "INACTIVE", "SOLD_OUT", "AUCTION_COMPLETED", "DELETED" ]
                  },
                  "productCondition" : {
                    "type" : "string",
                    "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                    "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                  },
                  "listingType" : {
                    "type" : "string",
                    "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                    "enum" : [ "NORMAL", "AUCTION" ]
                  },
                  "categoryId" : {
                    "type" : "integer",
                    "description" : "소속 카테고리 ID",
                    "format" : "int64"
                  },
                  "images" : {
                    "type" : "array",
                    "description" : "상품 이미지 목록",
                    "items" : {
                      "required" : [ "imageType", "imageUrl", "sortOrder" ],
                      "type" : "object",
                      "properties" : {
                        "imageUrl" : {
                          "type" : "string",
                          "description" : "상품 이미지 URL"
                        },
                        "imageType" : {
                          "type" : "string",
                          "description" : "{THUMBNAIL=THUMBNAIL, DETAIL=DETAIL}",
                          "enum" : [ "THUMBNAIL", "DETAIL" ]
                        },
                        "sortOrder" : {
                          "minimum" : 0,
                          "type" : "integer",
                          "description" : "이미지 노출 순서"
                        }
                      },
                      "description" : "상품 이미지 요청 DTO.\n\n<p>상품 등록 및 수정 요청에서 이미지 URL과 정렬 정보를 함께 전달할 때 사용합니다."
                    }
                  }
                },
                "description" : "수정할 상품 정보"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "수정된 상품 상세 정보",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "상품 ID",
                      "format" : "int64"
                    },
                    "name" : {
                      "type" : "string",
                      "description" : "상품명"
                    },
                    "description" : {
                      "type" : "string",
                      "description" : "상품 설명"
                    },
                    "price" : {
                      "type" : "number",
                      "description" : "상품 가격"
                    },
                    "productStatus" : {
                      "type" : "string",
                      "description" : "{ACTIVE=ACTIVE, INACTIVE=INACTIVE, SOLD_OUT=SOLD_OUT, AUCTION_COMPLETED=AUCTION_COMPLETED, DELETED=DELETED}",
                      "enum" : [ "ACTIVE", "INACTIVE", "SOLD_OUT", "AUCTION_COMPLETED", "DELETED" ]
                    },
                    "productCondition" : {
                      "type" : "string",
                      "description" : "{NEW=NEW, LIKE_NEW=LIKE_NEW, GOOD=GOOD, FAIR=FAIR, POOR=POOR}",
                      "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
                    },
                    "listingType" : {
                      "type" : "string",
                      "description" : "{NORMAL=NORMAL, AUCTION=AUCTION}",
                      "enum" : [ "NORMAL", "AUCTION" ]
                    },
                    "sellerNickname" : {
                      "type" : "string",
                      "description" : "판매자 닉네임"
                    },
                    "categoryId" : {
                      "type" : "integer",
                      "description" : "카테고리 ID",
                      "format" : "int64"
                    },
                    "categoryName" : {
                      "type" : "string",
                      "description" : "카테고리명"
                    },
                    "images" : {
                      "type" : "array",
                      "description" : "상품 이미지 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "이미지 ID",
                            "format" : "int64"
                          },
                          "imageUrl" : {
                            "type" : "string",
                            "description" : "상품 이미지 URL"
                          },
                          "imageType" : {
                            "type" : "string",
                            "description" : "{THUMBNAIL=THUMBNAIL, DETAIL=DETAIL}",
                            "enum" : [ "THUMBNAIL", "DETAIL" ]
                          },
                          "sortOrder" : {
                            "type" : "integer",
                            "description" : "이미지 노출 순서"
                          }
                        },
                        "description" : "상품 이미지 응답 DTO.\n\n<p>상품 상세 정보에 포함되는 개별 이미지 메타데이터를 나타냅니다."
                      }
                    },
                    "createdAt" : {
                      "type" : "string",
                      "description" : "상품 생성 일시"
                    },
                    "updatedAt" : {
                      "type" : "string",
                      "description" : "상품 수정 일시"
                    }
                  },
                  "description" : "수정된 상품 상세 정보"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "상품을 삭제합니다.",
        "description" : "상품을 삭제합니다.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "삭제할 상품 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "nickname",
          "in" : "header",
          "description" : "요청한 판매자 닉네임",
          "required" : true,
          "schema" : { }
        } ],
        "responses" : {
          "200" : {
            "description" : "No Content"
          }
        }
      }
    },
...
  },
...
}